Convert MTK assertions from methods on XCTestCase to functions in order to avoid the need to explicitly reference self when using them within escaping closures.
