+++ MetovaTestKit/ExceptionTesting/MTKExceptionTesting.swift
@@ -29,53 +29,50 @@
 
 import XCTest
 
-extension XCTestCase {
+/**
+ Synchronously tests the provided block for exceptions. If it would throw an exception, the exception is caught and the test fails. Otherwise, the test passes.
+ 
+ - parameter message:   The message to log upon test failure.
+ - parameter file:      The file the test is called from.
+ - parameter line:      The line number the test is called from.
+ - parameter testBlock: The block to test.
+ 
+ - returns: The caught exception. If no exception was thrown, returns `nil`.
+ 
+ - warning: You should not rely on `XCTestExpectation` fulfillment in this block. If an exception is thrown before fulfillment, the expectation will never be fulfilled. `XCTestExpectation` should be unnecessary as the block is executed synchronously.
+ 
+ - warning: This will only catch Objective-C-style exceptions. Swift's `fatalError`s are not caught by this test.
+ */
+@discardableResult public func MTKAssertNoException(message: @autoclosure () -> String? = nil, file: StaticString = #file, line: UInt = #line, testBlock: () -> Void) -> NSException? {
     
-    /**
-     Synchronously tests the provided block for exceptions. If it would throw an exception, the exception is caught and the test fails. Otherwise, the test passes.
-     
-     - parameter message:   The message to log upon test failure.
-     - parameter file:      The file the test is called from.
-     - parameter line:      The line number the test is called from.
-     - parameter testBlock: The block to test.
-     
-     - returns: The caught exception. If no exception was thrown, returns `nil`.
-     
-     - warning: You should not rely on `XCTestExpectation` fulfillment in this block. If an exception is thrown before fulfillment, the expectation will never be fulfilled. `XCTestExpectation` should be unnecessary as the block is executed synchronously.
-     
-     - warning: This will only catch Objective-C-style exceptions. Swift's `fatalError`s are not caught by this test.
-     */
-    @discardableResult public func MTKAssertNoException(message: @autoclosure () -> String? = nil, file: StaticString = #file, line: UInt = #line, testBlock: () -> Void) -> NSException? {
-        
-        let message = message() ?? "Caught exception while executing test block."
-        let exception = MTKCatchException(testBlock)
-        
-        XCTAssertNil(exception, message, file: file, line: line)
-        
-        return exception
-    }
+    let message = message() ?? "Caught exception while executing test block."
+    let exception = MTKCatchException(testBlock)
     
-    /**
-     Synchronously tests the provided block for exceptions. If it does not throw an exception, the test fails. If it would throw an exception, the exception is caught and the test passes. Any code below the thrown exception is not executed.
-     
-     - parameter message:   The message to log upon test failure.
-     - parameter file:      The file the test is called from.
-     - parameter line:      The line number the test is called from.
-     - parameter testBlock: The block to test.
-     
-     - returns: The caught exception. If no exception was thrown, returns `nil`.
-     
-     - warning: You should not rely on `XCTestExpectation` fulfillment in this block. If an exception is thrown before fulfillment, the expectation will never be fulfilled. `XCTestExpectation` should be unnecessary as the block is executed synchronously.
-     
-     - warning: This will only catch Objective-C-style exceptions. Swift's `fatalError`s are not caught by this test.
-     */
-    @discardableResult public func MTKAssertException(message: @autoclosure () -> String? = nil, file: StaticString = #file, line: UInt = #line, testBlock: () -> Void) -> NSException? {
-        
-        let message = message() ?? "Did not catch exception while executing test block."
-        let exception = MTKCatchException(testBlock)
-        
-        XCTAssertNotNil(exception, message, file: file, line: line)
-        
-        return exception
-    }
+    XCTAssertNil(exception, message, file: file, line: line)
+    
+    return exception
+}
+
+/**
+ Synchronously tests the provided block for exceptions. If it does not throw an exception, the test fails. If it would throw an exception, the exception is caught and the test passes. Any code below the thrown exception is not executed.
+ 
+ - parameter message:   The message to log upon test failure.
+ - parameter file:      The file the test is called from.
+ - parameter line:      The line number the test is called from.
+ - parameter testBlock: The block to test.
+ 
+ - returns: The caught exception. If no exception was thrown, returns `nil`.
+ 
+ - warning: You should not rely on `XCTestExpectation` fulfillment in this block. If an exception is thrown before fulfillment, the expectation will never be fulfilled. `XCTestExpectation` should be unnecessary as the block is executed synchronously.
+ 
+ - warning: This will only catch Objective-C-style exceptions. Swift's `fatalError`s are not caught by this test.
+ */
+@discardableResult public func MTKAssertException(message: @autoclosure () -> String? = nil, file: StaticString = #file, line: UInt = #line, testBlock: () -> Void) -> NSException? {
+    
+    let message = message() ?? "Did not catch exception while executing test block."
+    let exception = MTKCatchException(testBlock)
+    
+    XCTAssertNotNil(exception, message, file: file, line: line)
+    
+    return exception
 }
