+++ MetovaTestKitTests/FailureRecordingTests.swift
@@ -43,10 +43,10 @@ class FailureRecordingTests: MTKBaseTestCase {
     func testRecordingFailureWithMessageAndDescription() {
         
         let line: UInt = 100
-        let testFailureExpectation = TestFailureExpectation(description: "Description - Message", filePath: #file, lineNumber: line)
+        let testFailureExpectation = TestFailureExpectation(description: "failed - Description - Message", filePath: #file, lineNumber: line)
         
         expectTestFailure(testFailureExpectation) {
-            recordFailure(withMessage: "Message", description: "Description", inFile: #file, atLine: line, expected: true)
+            MTKRecordFailure(withMessage: "Message", description: "Description", file: #file, line: line)
         }
     }
 }
